docs(primitives): standardize primitives storybook
and group them under the primitives section on the docs
