+++ packages/primitives/src/clickable/stories/clickable.examples.stories.tsx
@@ -3,16 +3,10 @@ import React from 'react'
 import { Clickable, ClickableBubble } from '../index'
 
 export default {
-  title: 'primitives/clickable',
-}
-
-export function Default() {
-  return (
-    <Clickable onClick={() => alert('Parent click')}>
-      <p>Select me will NOT trigger clicks</p>
-      <button onClick={() => alert('Child click')}>Child</button>
-    </Clickable>
-  )
+  title: 'primitives/clickable/examples',
+  parameters: {
+    chromatic: { disableSnapshot: true },
+  },
 }
 
 export function ClickBubble() {
