+++ packages/primitives/src/combobox/combobox-list.tsx
@@ -2,6 +2,11 @@ import type { ComponentPropsWithoutRef } from 'react'
 import React, { forwardRef } from 'react'
 import { ComboboxList as List } from '@ariakit/react'
 
+/**
+ * Renders a list for the combobox
+ *
+ * @kind primitives
+ */
 export const ComboboxList = forwardRef<HTMLDivElement, ComboboxListProps>(
   function ComboboxList(props, ref) {
     const { asChild = false, children, ...otherProps } = props
