+++ packages/primitives/src/combobox/stories/combobox.examples.stories.tsx
@@ -9,25 +9,12 @@ import {
 } from '../index'
 
 export default {
-  title: 'primitives/combobox',
-}
-
-export function Default() {
-  return (
-    <ComboboxProvider>
-      <label>
-        Select a fruit
-        <ComboboxInput placeholder="e.g., Apple" />
-      </label>
-      <ComboboxPopover>
-        <ComboboxItem value="apple">🍎 Apple</ComboboxItem>
-        <ComboboxItem value="grape">🍇 Grape</ComboboxItem>
-        <ComboboxItem value="orange">🍊 Orange</ComboboxItem>
-        <ComboboxItem value="strawberry">🍓 Strawberry</ComboboxItem>
-        <ComboboxItem value="watermelon">🍉 Watermelon</ComboboxItem>
-      </ComboboxPopover>
-    </ComboboxProvider>
-  )
+  title: 'primitives/combobox/examples',
+  parameters: {
+    chormatic: {
+      disableSnapshot: true,
+    },
+  },
 }
 
 const fruitList = [
