+++ packages/primitives/src/select/stories/select.examples.stories.tsx
@@ -18,23 +18,12 @@ import {
 } from '../../combobox'
 
 export default {
-  title: 'primitives/select',
-}
-
-export function Default() {
-  const [value, setValue] = useState('')
-
-  return (
-    <SelectProvider value={value} setValue={setValue}>
-      <SelectTrigger>Select: {value}</SelectTrigger>
-      <SelectPopover>
-        <SelectItem value="apple">Apple</SelectItem>
-        <SelectItem value="google">Google</SelectItem>
-        <SelectItem value="microsoft">Microsoft</SelectItem>
-        <SelectItem value="amazon">Amazon</SelectItem>
-      </SelectPopover>
-    </SelectProvider>
-  )
+  title: 'primitives/select/examples',
+  parameters: {
+    chormatic: {
+      disableSnapshot: true,
+    },
+  },
 }
 
 export function Composition() {
