--- packages/primitives/src/virtual/tests/virtual.test.tsx
@@ -1,5 +1,0 @@
-import { describe, test } from '@vtex/shoreline-test-utils'
-
-describe('virtual', () => {
-  test.skip('renders', () => {})
-})
