+++ packages/primitives/src/visually-hidden/visually-hidden.tsx
@@ -4,6 +4,7 @@ import { VisuallyHidden as BaseVisuallyHidden } from '@ariakit/react'
 
 /**
  * Renders an element that is visually only to screen readers
+ * @kind primitives
  * @example
  * <a href="#">
  *  Learn more<VisuallyHidden> about brazilian beaches</VisuallyHidden>
